+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2.3.2 [unreleased]
+
+- Added `_all` and `default_fields` options
+
+## 2.3.1
+
+- Added support for `reindex(async: true)` for non-numeric primary keys
+- Added `conversions_term` option
+- Added support for passing fields to `suggest` option
+- Fixed `page_view_entries` for Kaminari
+
 ## 2.3.0
 
 - Fixed analyzer on dynamically mapped fields